projects/auto-sample-registration/requirements.txt
opendataloader-pdf[hybrid]
pdf2image
pytesseract
Pillow
gspread
google-auth
python-dotenv
